add srp for pages/course
